Simplify cloud provider controller manager setup
A better way to restrict the to-be-launced controllers is to
directly filter down the initFuncConstructors map that's passed into
NewCloudControllerManagerCommand. This allows some of the other calls
to the cloud provider API to be removed. This makes it compatible with
Kubernetes 1.27, which has some breaking changes in the removed API
calls.

Also remove the second worker from the inttest, as it wasn't used at
all.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>