Streamline kubectl integration test
The common test matrix wasn't checking for stderr contents so far.
Address this by capturing both stdout and stderr when executing the test
commands and forward those along with the execution error to the check
function. Add asserts for stderr contents to all the tests in the
matrix.

Moreover, the integration test environment wasn't completely isolated.
for example, the support-bundle plugin that's installed in the footloose
image was leaking into the test setup. Address this by setting up a
private PATH and use that when invoking the test commands. That way, the
plugin-list and symlink-warning test cases can be integrated into the
common test matrix as well.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>