Don't print kubectl usage on unknwon subcommands
The basic kubectl command will never accept any arguments. This will be
handled more or less automatically by Cobra when used as a root command.
When run as a subcommand, Cobra will instead pass all of the arguments
to the command's run methods, which will trigger some unexpected output.
Address this by specifying NoArgs for the kubectl command.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>