Use separate streams for pod exec
Using the same stream for both stdout and stderr causes empty results for even commands that are not failing and are expected to have output.
I believe this can happen as the stderr stream is empty, and thus get's quickly to EOF, it stops (or otherwise closes) the stream copying.

Signed-off-by: Jussi Nummelin <jnummelin@mirantis.com>