inttests: Use correct arch in footloose image
The downloaded binaries in the footloose Docker image only work on
amd64 currently. This is problematic when the inttests fail on non-amd64
arches, since the support bundle collection won't work and there's only
limited information available to debug the failures.

Use the automatic platform arg TARGETARCH when downloading binaries. Set
the TARGETARCH arg manually to the value of GOARCH for standard builds,
as it's only automatically set when using the BuildKit backend.

Add another make target `check-footloose-alpine-buildx` that's intended
to test all platforms at once using `docker buildx` when hacking on the
Dockerfile locally.

Update troubleshoot to v0.61.0 in order to unlock AArch32 binaries and
remove the etcd_arch build arg as it is redundant right now.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>