Let addons inttest respect the suite's context
The function wait.PollImmediate and its callbacks didn't short-circuit
on interrupted contexts. This lead to situations in which Go's test
deadline would cancel hanging inttests without letting enough time to
collect diagnostic data.

Use the wait.PollUntilContextCancel function along with a check if the
test context has been cancelled in the callback instead. This should
guarantee that the test will exit with a failure whenever the context
gets interrupted.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>