Clear error in Helm chart status after successful reconciliation
Whenever errors occur during the reconciliation of Helm charts, they are
recorded in the error field of the Chart resource. However, the error
status field was not cleared after a subsequent successful
reconciliation, which caused the Chart resource's status to erroneously
report the chart application as failed, even if it had recovered. To
address this, the error status should be cleared whenever there are no
errors.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>