Replace Slack with Lens Forums in the rest of the docs
The Slack workspace is deprecated and will be shutdown in the future.
The new place for the community is the Lens Forums. They are indexed
by Internet search engines and offer better discoverability than Slack,
which requires sign-in even if people just want to read along without
actively participating.

Also add the k0s blog on Medium to the social links in the footer.

See: 3738420ad ("Replace Slack with Lens Forums in README")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>