DNS fixes in docker entrypoint for both legacy and nft
Docker might use nf_tables instead of legacy iptables mode. This breaks
the k0s-in-docker scenario when k0s is used on a non-default bridge
networking. Fix this by simply trying to fix both modes. One of them
will probably be empty, so the sed will be a no-op.

Also update the references to the kind scripts where the fix is
originating from.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>