Use Docker iid files consequently in builds
Docker has the capability of writing out the image IDs to dedicated
files. This is easier and more integrated than doing it via shell
commands and output redirection.

Use the content of those iid files instead of the hard-coded image tags.
This has the advantage of being consistent, especially on development
machines when hacking on different k0s versions in different folders
in parallel.

Reuse the snippet used to clean up the iid files for the embedded
binaries also in the main k0s build.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>