Add GitHub workflows for OS testing
Introduce a callable GitHub workflow that leverages the Terraform
modules for OS testing to bring up a k0s cluster on AWS against which
Kubernetes conformance tests will be executed.

The workflow features parallel test execution to speed up test duration,
if desired. Note that test results collected in parallel are most likely
not accepted by the Kubernetes conformance test repo. If the tests
should be submitted there, it's advised to set the `e2e-concurrency-
level` to 1.

Moreover, provide a matrix workflow that bundles up the first one into
a matrix build, so that multiple OS and network provider configurations
may be tested using a single workflow, and a nightly trigger which
chooses a single OS/network provider combination to be run on a nightly
basis.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>