Bump k0sctl to v0.15.5
This addresses a probably SELinux related provisioning problem noticed
on Fedora CoreOS 38.20230826.20.0.

Not touching the release workflow, as that one is hard to test and I
don't want to introduce a regression.

Also add bash to the Alpine VMs in the OS tests to satisfy k0sctl's/
rig's new hard dependency on bash.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>