Lower disk space eviction limits for network conformance inttest
Recently, the network conformance inttest has been failing regularly.
This is because the sonobuoy pods are being evicted when the available
disk space on the GitHub runners falls below the default eviction
limits. Mitigate this by lowering the eviction limits.

This is duct-tape solution. It's unclear whether the reduced amount of
free disk space on the GitHub runners is due to changes made by GitHub
or to changes in k0s.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>