Fix log collection in cli inttest
The cli inttest does not utilize the inttest framework to start and stop
k0s. Instead, it manages these operations independently through the k0s
command line.

However, in cases of test failures, the inttest framework attempts to
collect log files from its default paths, which is incorrect when using
'k0s start' and 'k0s stop'. To resolve this issue, set the LaunchMode to
OpenRC so that the logs are collected appropriately.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>