Fix wrong controller port used for pre-generated tokens command
The port number of controller used in the custom CA cert doc's pre-generated tokens section is wrong. It's using `6443`, that's for `kube-apiserver`. But what we need here is `9443` actually (the k0s-api). Took me a while debugging to find out that k0s is trying to join the controller server with a wrong port

Signed-off-by: Fang-Pen Lin <hello@fangpenlin.com>