Fix image detection for kubelet metrics in inttests
When inttests are executed with images that contain hashes, the image
detection will fail. That's because kubelet won't include the tag,
but only the hash in its metrics.

Moreover, properly regex-escape the match string, pull out common logic
out of the retry loop and add the missing error check to the scanner.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>