Fix "executable in PATH" sysprobe
The loop over the varargs was not capturing the loop variable correctly,
so all the probes would check the same binary. Remove the loop and let
callers call the function multiple times for each executable. This is
less error prone. That bug was first introduced in 0f474b8f9, but didn't
become effective before 0f474b8f9.

Fixes: 0f474b8f9 ("Some new sysinfo probes")
Fixes: 19c800dfa ("Add sysprobe and some docs for required mount/umount")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>