Add functions using the types of values as context keys
Extend the existing k0scontext package with utility functions
simplifying context-based data management in a type-safe and ergonomic
manner. Each distinct type serves as a key for context values, removing
the necessity for an extra key constant. The type itself becomes the
key.

Add some comments for the already existing functions as well.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>