Remove necessity to suppress NotFound errors in Applier
Use a fresh resource builder for each stack application. The cached
resource builder kept around references to all the files that have been
added to it. Using a fresh one eliminates the need to suppress NotFound
errors, previously arising from attempts to load files no longer present
in the stack's folder.

Correct the corresponding integration test which used the wrong label
selector. The test mistakenly used the name of the chart resource
instead of the name of the helm release, so that the test always passed,
even if the uninstall didn't happen.

Also adjust the unit tests to better respond to error conditions and
remove some duplicate checks.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>