Assume Helm release has been uninstalled if not found
During the deletion of chart resources, it might happen that the Helm
release has already been uninstalled by a previous reconciliation loop,
or maybe manually by a cluster administrator.

Accomodate for that by treating release not found errors as a sign that
the release has already been uninstalled and continue with the chart
deletion process.

Add inttest for assuming non-existent releases as uninstalled.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>