Fix Code of Conduct linking in docs
There was a symlink that pulled in the Code of Conduct from the
repository root into the docs build. The link in the Code of Conduct was
supposed to be working from within the docs.

Cleanup that situation by removing one Code of Conduct file and update
the symlink. Also let changes to the Code of Conduct trigger the docs
build workflow, so that these errors might be caught earlier in the
future.

Fixes: 7aa222fa6 ("Update link to Code of Conduct document")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>