Propagate error when deleting stack
The error returned from `getPruneableResources` was not being propagated to the caller. This changelist fixes that by propagating the error.
The usage of `sync.WaitGroup` has been replaced with a context-aware `errgroup.Group` to allow failing fast by terminating other go-routines when one of the API server calls fails.
The usage of `sync.Mutex` has also been replaced with a buffered channel.

In the future, it might be worth considering bounding the number of go-routines spawned to avoid hitting API rate limits, and to turn the `prune` method into
a streaming pipeline that spawns separate go-routines for finding and deleting prunable resources, instead of listing all prunable resources first before starting the deletion.

Signed-off-by: cpu1 <patwal.chetan@gmail.com>