Add k0sctl version lookup to vars.sh
The vars.sh script is used to obtain variables from Makefile variables
in scripts. Even if the k0sctl version is not maintained in a Makefile,
but rather in a go.mod file, we can use the same interface for scripts.
This makes the origin transparent and we can special-case anything in
the vars.sh script, instead of introducing other ad-hoc mechanisms.

Moreover, include the k0sctl version in the build-env by default. It's
fast and safe enough to be included by default, even if not required.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>