Prove airgap bundle processing in AP inttest
The Autopilot airgap inttest is operating on an already bootstrapped
cluster. This means that all the images have already been pulled on the
worker as soon as the Autopilot update kicks in. Hence, the update will
succeed even if the airgap bundle functionality is broken.

To have an actual proof of correct processing, the inttest now checks
the pinning of the images. Before the update, none of them should be
pinned, after the update all of them.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>