Enable ipv6 for kube-router
kuber-router supports ipv6. Enable this and add dualstack test.

Unlike calico, kube-router requires that the node has a valid ipv6
InternalIP. This is only set with kublet when --node-ip is specified,
and this option requires that we also set the ipv4. Mimic the logic from
kubelet to find the ipv4 same way as kubelet.

https://github.com/cloudnativelabs/kube-router/blob/master/docs/ipv6.md#how-to-enable-dual-stack-functionality

Signed-off-by: Natanael Copa <ncopa@mirantis.com>