Remove unused bootstrap-api-auth token usage
It has been added as part of the initial Windows support,
but it seems it has never been in use. Supposedly this is a
remnant of an intermediary variant, and has been renamed to
bootstrap-api-worker-calls.

See: d2251a39d ("Implement different roles in the token/manager to distinguish api for controller nodes and api for worker nodes")
See: 55b52270e ("Apply lost in rebase commit for join token generation")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>