bug: retry saving helm chart object status
if we fail to save the status, specially after installing the chart, we
will try to install the chart again during the next reconcile (its
Status.Status.ReleaseName will be empty). this pr attempts to mitigate
this in two ways:

- before updating fetch the newest version of the chart.
- retry if it is still on conflict.

this may be overkill as it most likely won't fail after we fetch an
updated version of the chart but given the importance of this operation
we better be safe than sorry.

when this bug this solves happens this is what is reported in the chart
status:

```
can't install loadedChart ```chart-name```: cannot re-use a name that is
still in use
```

Signed-off-by: Ricardo Maraschini <ricardo.maraschini@gmail.com>