Rename CRIConfigurer to configurer
Now that the containerd Component lives in the same package as the
configurer, it can be made private. The factory function was already
named NewConfigurer. This struct specifically handles the merging of
CRI plugin sections, but as far as k0s is concerned, it handles all of
containerd's configuration. So it makes sense to strip the prefix from
its name.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>