[Docs] Fix k0s in docker documentation
This commit changes a few things

1- By default make k0s controller also a worker, IMO it's more logical
to do it the way the documentation is written.

2- Remove the mount to /sys/fs/cgroup as it's not really required. I
assume this was never required and comes from the 4.3.0  release notes.
It's probably still required for systemd but we don't run systemd.

3- Remove unnecessary tmpfs in /run/ and /var/run in docker-compose.
They are not wrong but not required either and this is consistent with
what we do in the regular docker run.

4- Always make cgroup: host. This works with both cgroups v1 and v2 but
not having this breaks cgroups v2.

Signed-off-by: Juan-Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>