Pass along containerd config as return value
Return the containerd config as a string instead of having the
configurer write it to disk. This reduces the number of struct fields
required by the configurer, and makes testing less dependent on the
file system.

Let the containerd component write the file instead.

Remove the hard-coded string literal for the paths to
containerd-cri.toml and use k0s' RunDir instead. This already takes into
account the differences between Windows and UNIX-like operating systems,
and also takes into account the different defaults when running as
non-root user.

Remove the escapedPath function and inline it directly before the
containerd component generates the template.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>