Improve k0s-managed containerd config detection
The previous detection method was to read the whole file and search each
line for the magic string "# k0s_managed=true". This seems excessive
and unexpected, as it will catch any substring, even if it is not at the
beginning of the file, the start of a line, or even inside a comment. It
would even have caught this as part of a TOML string value.

Change this behavior by expecting the first line of the file to be
exactly this magic string. All files generated by k0s will have exactly
that first line. Also return any previously swallowed errors back to
the caller.

Add more test cases to ensure that the generated default config is
recognized as being managed by k0s, and that a file not starting with
the magic line is recognized as not being managed by k0s.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>