Replace the Machine ID usage with an "Invocation ID"
Instead of relying on an externally maintained "machine identity"
that may or may not exist on hosts, use a so-called Invocation ID. The
Invocation ID is a randomly generated 32-byte hexadecimal string that
is created during process startup. The name is inspired by systemd's
INVOCATION_ID.

K0s doesn't need a unique host identifier that is stable across process
invocations. It only needs to have unique identifiers per process.
This removes a dependency on the k0s host machine (some machine-id
file or registry entry),  a Go module dependency, Machine ID related
documentation and system probes. It also makes the use of a unique ID
obvious. The machine ID has been hard-coded into the guts of the leader
election and telemetry components. The invocation ID is passed around
instead.

Stub out the use of the anonymous ID in telemetry, as the new ID is no
longer stable and would be useless for this type of use.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>