Make the bridge cleanup step optional based on OS
The bridge build step was a no-op on non-Linux systems. Instead of
adding and executing a no-op step, exclude it from the list of steps
altogether if it's not supported. Have a factory function that can
return a step or nil, depending on whether the bridge cleanup is
supported or not. Remove the dummy implementation entirely.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>