Ignore token-file argument when value not needed
Both "k0s controller" and "k0s worker" take either [TOKEN] as an
argument or a path to a file containing the join token via --token-file.

Before this PR, when --token-file was given, the file was read upon
spin-up, even when the node has joined already and won't utilize the
token. As the token is not needed after join, you would expect that you
can or even should remove the file, but that makes k0s refuse to start
unless the --token-file argument is removed from the systemd unit or
other startup scripts.

This PR makes the controller and worker read the contents of the file
only right before the value is getting used. This allows --token-file
argument to be ignored when it is not needed and the file not existing
will not generate an error.

Signed-off-by: Kimmo Lehto <klehto@mirantis.com>