Remove binary before building it
The Go compiler doesn't touch the binary if it doesn't detect any code
changes. (Presumably by checking the BuildID of the old and the new
binary?) However, the Makefile relied on a regeneration of the binary
in any case, so that it can reappend the bindata file agan. Mitigate
this by deleting the binary before the build. That way, the Go compiler
doesn't have a choice.

Fixes: 08df7f565 ("Provide empty asset offsets without codegen")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>
(cherry picked from commit ad98aac565fa99ad320ac417a776d48e3f67d782)