Structured URL replacement
Instead of doing a string replacement on the raw kubeconfig, parse it
and do a structured replacement of the URL. Adjust the error message
when the admin config is missing. Add some comments that give some
hints on how the kubeconfig admin subcommand relates to the kubeconfig
generation in certificates.go.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>