More helpful error message for empty/absent token files
An empty token file is a typical error condition when k0s gets installed
via k0sctl, but fails to join the cluster when first started. The
previous error message was a little bit cryptic. Try to be more helpful
by suggesting to create a new token or try to reinstall the node.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>