Use a versioned worker config stack
The worker config ConfigMaps have the Kubernetes version in their names.
This is to ensure a seamless upgrade path when upgrading minor k0s
versions. However, the RBAC resources and the stack name itself didn't
have the version in their names. This means that as soon as a controller
with the new minor k0s version takes over, it removes all previous
worker ConfigMaps from the cluster, effectively defeating the purpose of
the versioned names.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>