Change kine metrics port from 8080 to 2380
The default port is 8080. This conflicts with kube-router, which
since 2.x fails if it cannot bind to its metrics port. This leads to a
situation where the default configuration for a single node k0s cluster
is broken, as kube-router will go into a crash loop. Kube-router 1.x
will not be able to provide metrics, but will simply log the error and
move on. The integration tests did not cover this case properly, as
the kube-router manifests did not specify a readiness probe that the
inttests need to detect problems with daemon sets.

Port 2380 was chosen because it is normally used by etcd. Kine also uses
port 2379 for the client API, so why not use 2380 as well?

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>