Make k0s reset fail if it can't reach containerd
Prior to this commit, if the containerd unix socket wasn't listening
grpc.Dial would try to connect forever.

This commit establishes the connection in the background and the actual
call will fail if it has to.

Also we implement a single context for all the operations so that we can
cancel the execution with control c.

Co-authored-by: Tom Wieczorek <twz123@users.noreply.github.com>
Signed-off-by: Juan-Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>
(cherry picked from commit 4e7b82aa490911f9589dfe4d74c8024bddfba060)