feat: unpin images from unknown oci bundles
after importing all bundles from the oci directory we go through all
containerd images and unpin all images that were imported from bundles
that no longer exist on the system.

we start now to add the following label when importing:

- k0sproject.ocibundle.paths

on this label we keep a list of all oci bundles that contain such image
and we only unpin it if all the oci bundles were removed from the disk.

as now we are unpinning images based on their presence on disk we start
to operate also on rename and remove fs events.

this commit also address an issue where we could leave a containerd
connection opened.

Signed-off-by: Ricardo Maraschini <ricardo.maraschini@gmail.com>