Don't use the wrong PID in error strings
Instead of using the PID in the argument, the error strings were
constructed using the PID of the currently running process. This would
have caused a panic because the PID file cleanup is called before the
supervised process is started. So the s.cmd would have been nil.

Don't put the PID in the errors at this level at all, but wrap the error
once with the actual PID read from the PID file in maybeKillPidFile.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>