Whitelist local IPv6 CIDRs when airgapping nodes
This does the same to IPv6 that's done for IPv4. Allows the airgap
test to function if the SSH connection runs via IPv6. Also: check if
ip6table_filter is already loaded before trying to modprobe it. This
allows the integration test to work without sudo/doas when the module is
already loaded.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>