Use a ten second timeout for join requests
This prevents k0s from hanging on idle network connections, such as
random glitches, badly behaving load balancers, and so on.

Add a context to the JoinClient methods and call them with a context
that times out after 10 seconds when joining a controller. This has
the side effect that joining will also be interruptible by SIGTERM and
Ctrl+C.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>