Synchronize k0s config create code and docs
Convert the config to unstructured to be able to tweak the actual YAML.
Remove the creationTimestamp from the output. Remove the omitempty
flag from the NLLB enabled field, so that it's false value makes it
into the output. Remove all the parts from the docs that don't show up
in reality.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>