Fix testify.Suite Run() usages across tests
The suite retains its own testing.T pointer, which is used
for the suite's assertions. Whenever a sub-test is created via
suite.T().Run(...), the suite's internal pointer remains unchanged and
points to the parent test. Using the suite in the sub-test should then
be avoided to not run into errors like this:

> test executed panic(nil) or runtime.Goexit: subtest may have called
> FailNow on a parent test

An alternative is to use suite.Run() directly. This will adjust the test
pointer as long as the test func is running. This way, the suite remains
usable in the sub-test. Note that this only works for non-parallel
tests.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>