Install QEMU emulator in e2e ostests if needed
This allows for getting the k0s version from the binary if it's for a
different architecture. Another approach could be to pass on the k0s
version from the build job into the e2e job, but since there's already a
k0s-version variable with a special meaning for an empty value, there's
not a straight forward way to do this without making things even more
complicated.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>