Use dynamic base IDs for NLLB envoy process
Envoy uses some abstract UDP sockets with fixed names for its
hot-reloading facilities. Since the NLLB envoy process runs on the host
network, and abstract sockets are part of the network namespace, this
is a blocker for running multiple envoy processes in parallel without
tuning their startup parameters, so that they don't use the same socket
names and fail.

The --use-dynamic-base-id CLI flag makes envoy pick random base IDs
(the base ID is part of the abstract socket names) and retry for a while
until it can bind to a socket that is not already in use. This makes
NLLB work well with any other envoys that may be running on the k0s
worker's host machine.

Now that we are providing custom arguments to the image, we need to
duplicate the default ones, which is the reference to the configuration
file.

See: https://www.man7.org/linux/man-pages/man7/network_namespaces.7.html
See: https://www.envoyproxy.io/docs/envoy/v1.31.0/operations/cli.html#cmdoption-use-dynamic-base-id
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>