Allow reading config file from stdin
This was documented to be supported for a long time, but never really
worked, except for "k0s config validate".

When constructing the CfgVars struct, capture stdin, or the cobra
command's input if specified. When constructing the node config, check
if a hyphen was given as the startup path. If so, grab the previously
saved input reader and get the config from there.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>