Add explicit names to all controller-runtime controllers
New releases of controller-runtime will enforce unique names per
controller, as the names become part of logs and metrics. Duplicate
names will be very confusing. If no name is explicitly given,
controller-runtime will fallback to generate a name based on the
resource kind that's being managed by the controller, if any. Especially
Autopilot registers mutliple controllers for the same kind, hence it
will have overlapping controller names.

Add the "Named" call to all the places in which a controller is built.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>