tests: make sure we close before removing files
windows don't seem to be happy if we remove a file without closing it.
if you are seeing this commit then it means "yes, windows is not happy
about that". if you aren't seeing this comment then the answer is "no,
windows does not care about that" but you won't know (except if you
already knew).

Signed-off-by: Ricardo Maraschini <ricardo.maraschini@gmail.com>