Reintroduce K0smotronNetworks in tests
Custom docker networks were removed in #3602 because the custom networks
were revealing issues in our test suite.

This effectively means k0smotron cannot bump the k0s dependency to a
newer version.

This is a partial revert of 57b5560 getting only the minimum subset that
we require for k0smotron.

Signed-off-by: Juan-Luis de Sousa-Valadas Castaño <jvaladas@mirantis.com>