Introduce new leader election client
This is an alternative to the existing lease pool. Its main purpose
is to better coordinate callback invocations and avoid the need for
channels. Callers of the new client can still choose to send to channels
in the callbacks, if necessary.

The new client also attempts to address some concurrency issues with the
stock Kubernetes leader elector, which can call callbacks in unexpected
ways.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>