Let ClusterConfigInitializer apply the api-config stack
The k0s managed stacks don't have to live on disk. They can be used
directly from in-memory. Let the api-config stack be the first. Include
its application in the Start method of the ClusterConfigInitializer.
Have the applier manager ignore this stack and write an ignore note in
the stack directory.

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>